+++ src/pass/pass.entity.ts
@@ -76,10 +76,10 @@ export class Pass {
   toPlayerId!: number;
 
   @Field(type => Int)
-  @Column({ unsigned: true, type: 'tinyint' })
+  @Column({ unsigned: true, type: 'smallint' })
   x!: number;
 
   @Field(type => Int)
-  @Column({ unsigned: true, type: 'tinyint' })
+  @Column({ unsigned: true, type: 'smallint' })
   y!: number;
 }
