+++ src/shot/shot.entity.ts
@@ -91,10 +91,10 @@ export class Shot {
   fromTeamId!: number;
 
   @Field(type => Int)
-  @Column({ unsigned: true, type: 'tinyint' })
+  @Column({ unsigned: true, type: 'smallint' })
   x!: number;
 
   @Field(type => Int)
-  @Column({ unsigned: true, type: 'tinyint' })
+  @Column({ unsigned: true, type: 'smallint' })
   y!: number;
 }
